ziwei Stars 全數轉移成 kotlin

+++ src/test/java/destiny/core/chinese/ziwei/HouseSeqDefaultImplTest.kt
@@ -4,10 +4,12 @@
 package destiny.core.chinese.ziwei
 
 import destiny.core.chinese.ziwei.House.*
-import org.junit.Assert.*
-import org.junit.Test
 import org.slf4j.LoggerFactory
 import java.util.*
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertSame
 
 class HouseSeqDefaultImplTest {
 
@@ -56,23 +58,23 @@ class HouseSeqDefaultImplTest {
 
   @Test
   fun getAheadOf() {
-    assertEquals(0, impl.getAheadOf(命宮, 命宮).toLong())
-    assertEquals(1, impl.getAheadOf(兄弟, 命宮).toLong())
-    assertEquals(2, impl.getAheadOf(夫妻, 命宮).toLong())
-    assertEquals(3, impl.getAheadOf(子女, 命宮).toLong())
-    assertEquals(4, impl.getAheadOf(財帛, 命宮).toLong())
-    assertEquals(5, impl.getAheadOf(疾厄, 命宮).toLong())
-    assertEquals(6, impl.getAheadOf(遷移, 命宮).toLong())
-    assertEquals(7, impl.getAheadOf(交友, 命宮).toLong())
-    assertEquals(8, impl.getAheadOf(官祿, 命宮).toLong())
-    assertEquals(9, impl.getAheadOf(田宅, 命宮).toLong())
-    assertEquals(10, impl.getAheadOf(福德, 命宮).toLong())
-    assertEquals(11, impl.getAheadOf(父母, 命宮).toLong())
+    assertEquals(0, impl.getAheadOf(命宮, 命宮))
+    assertEquals(1, impl.getAheadOf(兄弟, 命宮))
+    assertEquals(2, impl.getAheadOf(夫妻, 命宮))
+    assertEquals(3, impl.getAheadOf(子女, 命宮))
+    assertEquals(4, impl.getAheadOf(財帛, 命宮))
+    assertEquals(5, impl.getAheadOf(疾厄, 命宮))
+    assertEquals(6, impl.getAheadOf(遷移, 命宮))
+    assertEquals(7, impl.getAheadOf(交友, 命宮))
+    assertEquals(8, impl.getAheadOf(官祿, 命宮))
+    assertEquals(9, impl.getAheadOf(田宅, 命宮))
+    assertEquals(10, impl.getAheadOf(福德, 命宮))
+    assertEquals(11, impl.getAheadOf(父母, 命宮))
 
-    assertEquals(-1, impl.getAheadOf(相貌, 命宮).toLong()) // 全書派，不存在 相貌宮
-    assertEquals(-1, impl.getAheadOf(命宮, 相貌).toLong()) // 全書派，不存在 相貌宮
-    assertEquals(-1, impl.getAheadOf(相貌, 兄弟).toLong()) // 全書派，不存在 相貌宮
-    assertEquals(-1, impl.getAheadOf(兄弟, 相貌).toLong()) // 全書派，不存在 相貌宮
+    assertEquals(-1, impl.getAheadOf(相貌, 命宮)) // 全書派，不存在 相貌宮
+    assertEquals(-1, impl.getAheadOf(命宮, 相貌)) // 全書派，不存在 相貌宮
+    assertEquals(-1, impl.getAheadOf(相貌, 兄弟)) // 全書派，不存在 相貌宮
+    assertEquals(-1, impl.getAheadOf(兄弟, 相貌)) // 全書派，不存在 相貌宮
   }
